+++ packages/raystack/components/calendar/index.tsx
@@ -1,5 +1,6 @@
 export type { DateRange } from 'react-day-picker';
 export type { CalendarProps, CalendarPropsExtended } from './calendar';
 export { Calendar } from './calendar';
+export type { DatePickerProps, DatePickerSlotProps } from './date-picker';
 export { DatePicker } from './date-picker';
 export { RangePicker } from './range-picker';