+++ packages/raystack/components/data-view-beta/data-view.types.tsx
@@ -6,6 +6,7 @@ import type {
   FilterTypes,
   FilterValueType
 } from '~/types/filters';
+import type { FilterChipCalendarProps } from '../filter-chip';
 import type { BaseSelectProps } from '../select/select-root';
 
 export type DataViewMode = 'client' | 'server';
@@ -80,6 +81,7 @@ export interface DataViewField<TData = any> {
   defaultFilterValue?: unknown;
   filterProps?: {
     select?: BaseSelectProps;
+    calendar?: FilterChipCalendarProps;
   };
 
   // ordering / grouping / visibility capability